+++ krita_plugin/krita_diff/__init__.py
@@ -1,10 +1,10 @@
 from krita import DockWidgetFactory, DockWidgetFactoryBase, Krita
 
+from .docker import *
 from .hotkeys import Hotkeys
-from .krita_diff_ui import *
 
 instance = Krita.instance()
 instance.addExtension(Hotkeys(instance))
 instance.addDockWidgetFactory(
-    DockWidgetFactory("krita_diff", DockWidgetFactoryBase.DockLeft, KritaSDPluginDocker)
+    DockWidgetFactory("krita_diff", DockWidgetFactoryBase.DockLeft, SDPluginDocker)
 )
